up: added ambient light contribution

+++ pathtracer/src/color/Colors.hh
@@ -3,12 +3,12 @@
 #include "../scene_elements/fixed_size_vectors/Vector3D.hh"
 
 struct Colors {
-    static const Vector3D<uint8_t> RED;
-    static const Vector3D<uint8_t> GREEN;
-    static const Vector3D<uint8_t> BLUE;
-    static const Vector3D<uint8_t> YELLOW;
-    static const Vector3D<uint8_t> CYAN;
-    static const Vector3D<uint8_t> MAGENTA;
+    static const Vector3D<float> RED;
+    static const Vector3D<float> GREEN;
+    static const Vector3D<float> BLUE;
+    static const Vector3D<float> YELLOW;
+    static const Vector3D<float> CYAN;
+    static const Vector3D<float> MAGENTA;
 };
