+++ pathtracer/src/scene_elements/serialized/lights/Light.cc
@@ -1,10 +1,10 @@
 #include "Light.hh"
 
 Light::Light()
-        : color_(Vector3D<uint8_t>())
+        : color_(Vector3D<float>())
           , intensity_(1) {}
 
-Light::Light(float intensity, const Vector3D<uint8_t> &color)
+Light::Light(float intensity, const Vector3D<float> &color)
         : color_(color)
           , intensity_(intensity) {}
 
@@ -16,11 +16,11 @@ void Light::setIntensity_(float intensity_) {
     Light::intensity_ = intensity_;
 }
 
-const Vector3D<uint8_t> &Light::getColor_() const {
+const Vector3D<float> &Light::getColor_() const {
     return color_;
 }
 
-void Light::setColor_(const Vector3D<uint8_t> &color_) {
+void Light::setColor_(const Vector3D<float> &color_) {
     Light::color_ = color_;
 }
