+++ pathtracer/src/scene_elements/serialized/lights/PointLight.hh
@@ -8,7 +8,7 @@ class PointLight : public Light {
 public:
     PointLight(const Vector3D<float> &position, float radius);
 
-    PointLight(float intensity, const Vector3D<uint8_t> &color,
+    PointLight(float intensity, const Vector3D<float> &color,
                const Vector3D<float> &position, float radius);
 
     PointLight() = default;
